requirements.txt
flask
chromadb
requests
google-generativeai
python-dotenv
python-dateutil
